Remove dist/, add static to src/, Webpack build
Elm.js is a build artifact, so it's best to keep it out of
version control; same goes for dist/. The static files were
moved to src/static and are now copied to dist/ as they are with
Webpack. Webpack is now also responsible for compiling Elm and
adding the result to inline to dist/index.js. The result of
the build process is largely the same as before, but without
tying it to version control. Webpack allows for faster development
with yarn run dev.